Add files via upload
The dataset was collected from the paper “5G enabled contactless multi user presence and activity detection for independent assisted living” [4]. The authors introduced 5G for multi-user sensing applications. It is applicable for contactless monitoring and designed as an RF-sensing system that operates in the 5G frequency band, particularly at 3.75 GHz operating frequency. The dataset was divided into 16 classes, that is, a total of 1777 data samples/files, each represents a particular number of subjects and activities. The main data folder is subdivided into 16 folders corresponding to the 16 classes.